Vite Setup
feat: Created react project with 'npm create vite@latest . -- --template react-ts'
feat: Added .gitignore

I was unable to select TypeScript version from this project because it is nested inside the 'client' subfolder. As a result I have added a settings.json file with a link to typescript in the client subfolder to use the version specified there.